Use custom useQuery and useMutation hooks in TreeData and children
Support creating useMutation hooks that use either POST or PUT, and can pass in either one string or an array.
Create useTreeDataMutation and useTreeHistoryMutation hooks.
Fix createAPIFetch() to handle query and data being passed in an array, as they are currently when getPost() is used directly.
Return early with null if there's no treeData, instead of having everything hang off of a check on idTree.
Arrange props in consistent order.