Split the code into chunks
Load the /data page's code on demand, so that the user doesn't incur the cost of the full tree data when they're just viewing the map.
Re-enable source-map-loader, while ignoring react-router.
Clean up the Loading element and add a visible icon.
Don't load app.scss in Sidebar.scss, and let the App component include it.