Add side panel to show selected tree details or new tree controls
Add `/map` route to access the new behavior.
Add MapLayout component to manage the map and tree details.
Add Map component to render the mapbox and controls.
Show a selection state for the current tree on a separate layer.
Update Address section in new tree panel as the marker is dragged around.
Save changed notes to the backend after a short delay or on loss of focus.
Create useNewTree() hook to handle NewTreePanel interactions via a context.
Add MapboxControlPortal to wrap a Mapbox control created with React, and use it with the Plant and Adopt buttons.
Show the adoption directions in a dialog instead of a drawer from the left or inline in the details drawer.
Add ErrorBoundary.js to show an alert at the bottom of the screen.  Closing the alert will reload the page.
Set react-query to not refresh the queries when the window is refocused.
Refactor createUseQuery() to take an options param and enable a query to modify the request before it's passed to useQuery(), so that it can filter out certain requests to avoid server errors, like /trees?currentTreeId=null returning a 500.
Create Like, Liked, Adopt, Adopted checkboxes with appropriate fill colors.
Move maintenance buttons and notes field into a dialog.
Move Edit link to same row as like and adopt.
Add Map.css to include mapbox style overrides
Add `npm run docker` script to start docker-compose.
Add use-debounce and react-error-boundary packages.