Use `setQueryData()` to cache the data from the vector tiles
When the TreeDetailsPanel does a tree query through react-query, it will get this cached data immediately, and then will update when the server response comes back.
In CoreData, check that datePlanted exists before trying to format it, since date-fns will throw an exception otherwise.
Clean up Health to better handle invalid healthNum data and to keep the slider and health string display in sync.
Return null in Info if there is no info to display to avoid rendering a 0 when there isn't.
Add getNormalizedValue() and getValueByName() to treeHealth.js.
Move planting state out of NewTree and into newTreeState.isPlanting, so calling endPlanting() will automatically cancel the NewTreePanel.
Change startDrag() to beginDrag(), to be parallel with beginPlanting().