Ensure popup closes when mouse leaves a tree circle
Upgrade to mapbox-gl v2.7.0 so we can supply an array of layer IDs when adding the mouseleave handler.  That way, we can use a single handler for all the layers while still getting a mouseleave when the mouse moves over a popup.
Fix the order of calls in map.onload to ensure the public.treedata source gets added before the map layers get rendered.
Don't render any Map children until the map is loaded, and render MapLayers first.
Move the container in the portal components into state, so that when the container is created in response to an onAdd() call from the map, a re-render will be triggered.