Make add and tracker buttons both icons on a circle background
Use the add icon for the add button.
Use the POI marker for the tracker icon, and make it white when selected.
Create Crosshairs.js for the icon in the middle of the planting marker.
Move buttons to MarkerButtons.js.
Make the MapLayout background color the same as the map's land color.
Remove PlantTarget.js and TrackingToggle.js, which are no longer used.
Wrap the marker buttons and target in tooltips.
Add delay before showing all tooltips and set their max-width to 150.
Add tooltipSx prop to make it easy to customize the tooltip element.
Don't let the mouse hover over the tooltip.