Clean up from side panel PRs
Convert MapboxMarkerPortal to functional component.
Lazy-load the /map route.
Wrap the app in an ErrorBoundary component.
Fade in the loading page content so there's not a flicker of the tree image before the /map route loads.
Make handling of edited notes work like the health slider and fix a bug where edited notes didn't show after a reload.
Create useIsMobile hook using media query.
Update docs and vars in queries.js.
Make the marker buttons darker and add a white border to the outside it shows up better on a complicated background.
Suppress mouse events on the buttons so the map doesn't hear them and show tooltips.
Show a yellow dot in the middle of the crosshairs.
Clean up TreeDetailsPanel and CoreData.
Export a marker offset value from Marker.js so NewTree can adjust where the tree is planted based on the marker location.
Add postinstall script to run build:data so that the JS files are created in src/data/dist when the repo is created.
Add build timestamp to bottom of index.html.