Create React GeolocateControl component
Create UserLocationProvider and wrap MapLayout in it.
Create useUserLocation hook to share context between GeolocateControl and NewTree, and centralize access to the browser geolocation API.
When following the user, hide the geolocation marker and animate the planting marker.
In the GeolocateControl, show a blue icon with a dot when tracking the location and not planting, or when the planting marker is following the user.  If tracking is on but following is off, remove the center dot.
Default tracking on when tapping Plant on mobile.
Make planting marker grey instead of blue and make crosshairs yellow.
Add z-index to tooltips so they appear above the marker.
Refactor marker code and files.