feature/clustering
Added more clustering layers to the map. There are now two data sources
and three clustering layers. The geojson layer is so that we can cluster
overlapping cities. Some of the layers have been renamed and cleaned up.

The MapCluster layer is the outer zoom layer, with clusters for cities.
This is the geojson data source so we can cluster overlapping cities.
There is an issue with aggregating city names in here. That is a mapbox
issue that has been a long-standing issue for years.

The TreeCluster layer is for the middle zooms. It adds basic clusters of
trees.