Issue #266 Part 1: Update React 17->18 (#592)
* chore: update to React 18

I want to use this react-spring-carousel
component, which is dependent on React 18.
I update react and react-dom. To do the
update, I had to uninstall the deprecated
@mui/styles library.

* fix: update snapshots

I don't think these snapshots have anything
to do with the React update or removal
of @mui/styles. The changes have to with
CSS styles produced by Material UI
components, which must be a side effect
of a previous update to MUI which is not
in our control.

* chore: update npm packages

I saw Dependabot was giving
some errors about security
vulnerabilities in outdated
packages, so I updated everything
(within its semver range) using
`npm update`.

* Revert "fix: update snapshots"

This reverts commit 1c263a7385cbd27ff745fea8732d00e55fc30446.

* fix: update Tree.test.js snapshot

This snapshot is failing due to MUI changes,
but it needs to be updated for CI to pass.