test: Bring accessibility up to 100% in Lighthouse check (#606)
* fix: resolve accessibility issues

Fix the errors reported by Lighthouse.

* test: Add Lighthouse CI locally

Added the Lighthouse CI config file.
I had to tweak the info message
server/index.js emits, because Lighthouse
needs to see something like "Listening on..."
to detect that the server has started.

* ci: Added Lighthouse CI github action

Lighthouse CI should now run in
pre-flight. The scores for accessibility
and best practices need to be 100 for CI
to pass.

* test: Update snapshots

Adding a title for accessibility to Header
changed its test snapshot, so it had to be
updated.

* fix: pull the right @lhci/cli version

Pull the latest version of @lhci/cli
at major version 11.

* fix: install all npm dependencies

Some of the dev dependencies are needed
for the webpack build, so let's just
install everything.

* fix: install chromium in github action

Lighthouse needs chromium, but trying
to test its installation in a Github
action is exceedingly difficult even
with nektos/act. Let's see if this works.

* test: Use a docker image for chrome

Pulls in browserless/chrome docker image
to have chrome pre-installed. I'm having
a lot of trouble with the PROTOCOL_TIMEOUT
error during lhci autorun, which is still
unresolved by Lighthouse (it's the #1
issue on the GoogleChrome/lighthouse
repo).

* test: just download chrome manually

I do not want to deal with git auth
in docker images.

* test: Add Lighthouse token

Added the Lighthouse token so we
can see status checks. The
lighthouse github action I pulled
in has some issue with CHROME_PATH,
so I'm going back to running
lhci autorun by itself.

* fix: remove disable-storage-reset

reviewify recommended I remove it,
so I did. The machines tell us
what to do now.

* test: Use latest github actions

---------

Co-authored-by: Rose <rommims@gmail.com>